Fix: resolve remaining -Woverloaded-virtual warnings in our code (170 warnings)
- CriteriaParent: using CriteriaNode::unhook (derived overloads unhook(CriteriaNode*))
- QuestionDialog: using ModalDialog::init (overloads init(string) hides init(void))
- Box: using Mesh::ProcessDrawQueue (overloads with int drawqueue)
- GameCockpit: using Cockpit::Init (overloads Init(file) hides Init(file,bool))
- Removed 3 dead no-arg base Cockpit methods (DrawTargetBox, DrawTurretTargetBoxes,
  DrawTacticalTargetBox) — never called; the derived Sensor-arg versions are the real API
- SphereMesh/CityLights: changed ProcessDrawQueue(int,...) -> size_t to match the
  base Mesh signature (intended overrides, now true overrides)

All behavior-preserving. Eliminates the final 170 -Woverloaded-virtual warnings;
total warnings now 36 (from 5271).

+++ src/gfx/sphere.cpp
@@ -20,7 +20,7 @@ static GFXColor getSphereColor()
     return tmp;
 }
 
-void SphereMesh::ProcessDrawQueue( int whichpass, int whichdrawqueue, bool zsort, const QVector &sortctr )
+void SphereMesh::ProcessDrawQueue( size_t whichpass, int whichdrawqueue, bool zsort, const QVector &sortctr )
 {
     static GFXColor spherecol( getSphereColor() );
     GFXColor tmpcol( 0, 0, 0, 1 );
@@ -36,7 +36,7 @@ void SphereMesh::SelectCullFace( int whichdrawqueue )
     GFXEnable( CULLFACE );
 }
 
-void CityLights::ProcessDrawQueue( int whichpass, int whichdrawqueue, bool zsort, const QVector &sortctr )
+void CityLights::ProcessDrawQueue( size_t whichpass, int whichdrawqueue, bool zsort, const QVector &sortctr )
 {
     const GFXColor citycol( 1, 1, 1, 1 );
     GFXColor tmpcol( 0, 0, 0, 1 );
